requirements.txt
pdfplumber>=0.10.0
Flask>=2.0.0
Pillow>=9.0.0
pytesseract>=0.3.10
